commands: move version command to separate pkg
