+++ command/version/version_test.go
@@ -1,0 +1,15 @@
+package version
+
+import (
+	"strings"
+	"testing"
+
+	"github.com/mitchellh/cli"
+)
+
+func TestVersionCommand_noTabs(t *testing.T) {
+	t.Parallel()
+	if strings.ContainsRune(New(cli.NewMockUi(), "").Help(), '\t') {
+		t.Fatal("usage has tabs")
+	}
+}
